moving modal dialog to mercury/components
